Tweak description for SlintComponentGallery demo, font sizes and weights
This is another WIP tweaking some of the texts and text layout on the UI.
This was done during a conversation with Tobias in RocketChat on 2025-01-31
to see how much text we can fit with an increased font size.
This is mainly for Benny to see where further updates to the texts and
font layout shall be done.

Change-Id: I968f7b135839401b74ecda8f527ee7c019fff0a9